requirements.txt
annotated-types==0.6.0
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
docutils==0.20.1
ffpyplayer==4.5.1
filelock==3.12.4
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gradio_client==0.6.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.18.0
idna==3.4
jedi==0.19.1
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-examples==2.2.1
Kivy-Garden==0.1.5
packaging==23.2
parso==0.8.3
Pillow==10.1.0
plyer==2.1.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
python-dotenv==1.0.0
pywin32==306
PyYAML==6.0.1
replicate==0.15.4
requests==2.31.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
watchdog==3.0.0
websockets==11.0.3
